Added the generated Enforcer.doc to make it easier to read
The Enforcer.doc is the Amiga AutoDoc generated documentation from the
source code.  This is the same AutoDoc we used to generate the core
documentation for the AmigaOS and the ROM Kernel Manuals.  Our source
code had all of the documentation for the OS API built with this
so I used that same tool to build the documentation for Enforcer.

Since that tool is not included here (I do not have the rights to
distribute the tool) I figured I should store a generated version
of the documents.

Also put an image of my Amiga's Workbench display of the Enforcer
source directory.  Note that most source files have no icon defined
so they did not show up.  (Since web browsers don't know how to
render the *.icon files)