Add a note about text file encodings
The text files from the Amiga (including source code) were in ISO/ECMA Latin-1 codepage - an 8-bit character set that predates UNICODE and/or UTF-8.  As such, if looked at with UTF-8 editors/displays will make them look a bit strange.  However, 99.9%+ of the characters are in the ASCII space so it really does not matter.